pyproject.toml
[tool.poetry]
name = "chatgpt-telegram-bot"
version = "0.1.0"
description = ""
authors = ["nzooherd <nzooherd@gmail.com>"]
readme = "README.md"
packages = [{include = "chatgpt_telegram_bot"}]

[tool.poetry.dependencies]
python = "^3.8"
pydub = "^0.25.1"
telethon = "^1.28.2"
socks = "^0"
python-dotenv = {version = "^1.0.0", source = "pypi"}
openai = {version = "^0.27.4", source = "pypi"}
multiprocess = "^0.70.14"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

